Add beginnings of git sync
The git sync system at the moment is very bare-bones. In future, there
should be a channel for the "sync status" of the repo that is set to
True by git. Update and set to false on the creation of a new post,
editing of an existing post, or the publishing of a post.